requirements.txt
# Core ML dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.21.0
evaluate>=0.4.0

# Whisper specific
openai-whisper

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
webrtcvad>=2.0.10

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Text processing
jiwer>=3.0.0
editdistance>=0.6.0
langdetect>=1.0.9
sacrebleu>=2.3.0

# API and serving
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0
redis>=4.6.0
celery>=5.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
aiosqlite>=0.19.0

# Cloud and storage
boto3>=1.28.0
minio>=7.1.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk>=1.29.0

# Distributed training
deepspeed>=0.10.0
wandb>=0.15.0
tensorboard>=2.13.0

# Model optimization
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
optimum>=1.12.0

# Telegram bot
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Infrastructure
kubernetes>=27.2.0
docker>=6.1.0
pyyaml>=6.0

# Performance profiling
py-spy>=0.3.14
memory-profiler>=0.61.0
line-profiler>=4.1.0